support unicode, add cli help, improve slide layout
Unicode support required decoding the base64 string to binary,
then to bytes in an array, then to text with a decoder. This is
because Javascript is... an unfortunate language.

Adding command line help was trivial.

To improve the slide layout a new content div was necessary,
which simplified the CSS and ensures top line alignment for
quick reading across slides.

There is also an internal refactor here where the js and css
has been moved out to their own files. This makes the code
easier to reason about.